requirements.txt
# This is an implicit value, here for clarity
--index-url https://pypi.python.org/simple/


# for reqparse
six==1.9.0

# for ipaddr validator
netaddr==0.7.13

Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-Security==1.7.4
Flask-RESTful==0.3.2
py-bcrypt==0.4
pytz==2014.10
nose==1.3.4
aniso8601==0.92

#for MySQLdb
pymysql==0.7.2

eve==0.5.2

#apns
apns==2.0.1

# for check url resource exist
httplib2==0.9.2

# for img resize
Pillow==3.2.0

# for OAuth2
#apiclient==1.0.3
#google-api-python-client==1.5.1
rauth==0.7.2

pushjack==1.2.1
